Simplify dumb_update(), support arguments in smart_update()*
This change set simplifies and improves the feature requested in
issue #3 and implemented in pull request #4 by switching from an
optional list argument to `star-args' and applying the same calling
convention to smart_update() as well.

This is backwards incompatible with the implementation in pull
request #4 (which I merged into the `dev' branch but never published
to PyPI) and it's also technically backwards incompatible in the
sense that keyword arguments could previously be given to
smart_update() as positional arguments. When I publish this I will
bump the major version number.