Use ubuntu.com/mirrors.txt without trusting it 😇
Fixes #6. Rationale:

The feedback in issue #6 suggests that mirrors.ubuntu.com/mirrors.txt is
working properly again and should be preferred over scraping Launchpad.
However I prefer for apt-mirror-updater to be reliable "do what I mean"
software and mirrors.ubuntu.com/mirrors.txt has proven to be unreliable
in the past. So what can we do? Well, let's see:

- I think Launchpad is more reliable.
- Issue #6 suggests that mirrors.ubuntu.com/mirrors.txt is faster.

I think the best thing we can do is to change the Ubuntu mirror
discovery as follows:

1. Discover Ubuntu mirrors on Launchpad.
2. Try to discover mirrors using mirrors.ubuntu.com/mirrors.txt and iff
   successful, narrow down the list produced in step 1 based on the URLs
   reported in step 2.
3. Rank the discovered / narrowed down mirrors and pick the best one.

The reason why I'm adding this kind of complexity is because it has
bothered me in the past that Ubuntu mirror discovery was slow. This
helps a lot.